% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRReads}
\alias{miRReads}
\title{miRNA read counts}
\format{miRNA contains unprocessed miRNA read counts (\code{miRNA$raw.count}) 
and column data (\code{miRNA$col.dat}) that can be used for analysis.}
\usage{
miRReads
}
\description{
Full microRNA sequencing data of the pig (Sus scrofa) ovarian follicle 
used in Pan, Toms, et al. (2020).
}
\keyword{datasets}
