% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snoRReads}
\alias{snoRReads}
\title{snoRNA read counts}
\format{snoRNA contains unprocessed snoRNA read counts (\code{snoRNA$raw.count}) 
and column data (\code{snoRNA$col.dat}) that can be used for analysis.}
\usage{
snoRReads
}
\description{
Full small nucleolar RNA sequencing data of the pig (Sus scrofa) ovarian follicle 
used in Pan, Toms, et al. (2020).
}
\keyword{datasets}
